Move tests related to logging to their own module
