+++ tests/test_cases/test_cocotb/Makefile
@@ -43,6 +43,7 @@ MODULE := "\
 	test_edge_triggers,\
 	test_async_coroutines,\
 	test_handle,\
+	test_logging,\
 	"
 
 ifeq ($(shell python -c "import sys; print(sys.version_info >= (3, 6))"), "True")
